requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow-cpu==2.10.0
protobuf==3.20.3
ml-dtypes==0.2.0
imbalanced-learn
streamlit
